Correcting cgroup refactoring
