+++ azurelinuxagent/common/cgroups/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Microsoft Corporation
+# Copyright 2019 Microsoft Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
