+++ tests/common/test_cgroups.py
@@ -17,8 +17,10 @@
 
 from __future__ import print_function
 
-from azurelinuxagent.common.cgroups.cgroups import CGroupsTelemetry, CGroups, CGroupsLimits, \
-    CGroupsException, CGroupsLimits, BASE_CGROUPS, Cpu, Memory, DEFAULT_MEM_LIMIT_MIN_MB
+from azurelinuxagent.common.cgroups.cgroups import CGroupsTelemetry, CGroups
+
+from azurelinuxagent.common.cgroups.cgutils import CGroupsException, Cpu, Memory, CGroupsLimits, \
+    BASE_CGROUPS, DEFAULT_MEM_LIMIT_MIN_MB
 from azurelinuxagent.common.version import AGENT_NAME
 from tests.tools import *
